Create sed.c
User program “sed” in XV6, which reads the content from the pipe redirection or from the file. “sed” also accepts the string_replaced and the string_replaced_with, and these two strings are of the same arbitrary length. “sed” scans the inputs, finds all the exact matches of the string_relaced, and replaces them with the string_replaced_with. Finally, ‘sed’ prints the result to stdout.
Format with the file as an input: sed -ToBeReplaced -ToBeReplacedWith Filename
Format with the pipe as an input: “some process” | sed -ToBeReplaced -ToBeReplacedWith